Fix potential race condition in cipher usage
